+++ CHANGELOG.md
@@ -5,3 +5,5 @@ All notable changes to `make-action` will be documented in this file.
 ## 1.0.0 - 2025-07-21
 
 - Updated package namespace from `GlenBangkila\MakeAction` to `Webteractive\MakeAction`.
+- Added Laravel Prompts support for interactive command input.
+- Added `laravel/prompts` as a dev dependency.